genre-classifier-service/requirements.txt
absl-py==2.3.1
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
async-timeout==5.0.1
audioop-lts==0.2.2
audioread==3.0.1
aws-msk-iam-sasl-signer-python==1.0.2
boto3==1.35.59
botocore==1.35.99
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
idna==3.10
jmespath==1.0.1
joblib==1.5.2
keras==3.11.3
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
namex==0.1.0
numba==0.61.2
numpy==2.2.6
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.1.0
s3transfer==0.10.4
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
